Se arregla nombres de cardinales

+++ src/mapa/tipos/Precipicio.cpp
@@ -16,7 +16,7 @@ char Precipicio::obtener_tipo(){
 }
 int Precipicio::obtener_energia(int tipo_personaje){
 
-    if(tipo_personaje>=0 && tipo_personaje<=MAX_TIPO_PERSONAJE)
+    if(tipo_personaje>=0 && tipo_personaje<= (int)MAX_TIPO_PERSONAJE)
 
         return ENERGIAS[tipo_personaje][PRECIPICIO];
